// src/app/models/poitems.model.ts
export class PoItems {
    id: string;
    orderNumber: string;
    orderStatus: string;
    supplierName: string;
    productCategory: string;
    productBrand: string;
    productName: string;
    orderQuantity: Number;
    deliveryDate: Date;
    productPrice: Number;
    productsOrderPrice: Number;
}